Генерация обучающего yaml + рефакторинг

+++ ann_manager.py
@@ -15,6 +15,8 @@ import json
 import random
 import cv2
 
+from ultralytics import YOLO
+
 def update_lists():
     global videolist, processed_list
 
@@ -54,15 +56,23 @@ def view_video():
     if len(selection):
         subprocess.run(["python3", "annxml_view.py", processed_subfolder / processed_listbox.get(selection)])
 
-def process_video_item(item):
-    print('Process ', item)
+def process_video_and_annotation_item(item):
+    print('Process video and annotation ', item)
     annotation = Annotation()
     annotation.load(item)
     cap = cv2.VideoCapture(annotation.videofilename)
-    root_images_path = Path(project_folder_entry.get()) / "images"
-    images_path = root_images_path / Path(item).stem
-    rmtree(images_path, ignore_errors=True)
+    root_images_path = Path(project_folder_entry.get())
+    
+    images_path = root_images_path / "images" / Path(item).stem
+    labels_path = root_images_path / labels_folder_entry.get() / Path(item).stem
+
+    #rmtree(images_path, ignore_errors=True)
     images_path.mkdir(parents=True, exist_ok=True)
+
+    rmtree(labels_path, ignore_errors=True)
+    labels_path.mkdir(parents=True, exist_ok=True)
+
+    images = []
     while True:
         frame_number = int(cap.get(cv2.CAP_PROP_POS_FRAMES))
         ret, frame = cap.read()
@@ -70,22 +80,11 @@ def process_video_item(item):
             break
             
         framejpg = f'{images_path}/{frame_number}.jpg'
-        cv2.imwrite(framejpg, frame)
-    print(item, ' processed')
+        if not Path(framejpg).is_file():
+            cv2.imwrite(framejpg, frame)
 
-def process_annotation(item):
-    print('Process ', item)
+        frame_pos = annotation.get_frame_poses(frame_number)
 
-    labels_path = Path(project_folder_entry.get()) / "labels" / Path(item).stem
-    images_path = Path(project_folder_entry.get()) / "images" / Path(item).stem
-
-    rmtree(labels_path, ignore_errors=True)
-    labels_path.mkdir(parents=True, exist_ok=True)
-
-    annotation = Annotation()
-    annotation.load(item)
-    images = []
-    for frame_number, frame_pos in enumerate(annotation.frame_poses):
         frame_annotations=list()
         for pos in frame_pos:
             if pos.is_outside:
@@ -108,18 +107,14 @@ def process_annotation(item):
                     for el in ann_item:
                         print(el, file=f, end=' ')
                     print(file=f)
+    print(item, ' processed')
 
-    print(item, f' processed images: {len(images)}')
     return images
 
-def generate_images():
+def generate_dataset():
     with multiprocessing.Pool(workers_number) as pool:
-            pool.map(process_video_item, videolist)
+        all_images = pool.map(process_video_and_annotation_item, processed_list)
 
-def generate_labels():
-    with multiprocessing.Pool(workers_number) as pool:
-        all_images = pool.map(process_annotation, processed_list)
-    
     all_images = [item for image_item in all_images for item in image_item]
 
     nfold = 5
@@ -135,6 +130,13 @@ def generate_labels():
     ftr.close()
     fte.close()
 
+    model = YOLO(ann_manager_settings['yolo_nn_name'])
+
+    train_data = {"path": project_folder_entry.get(), "train": "train.txt", "val": "validation.txt", "names": model.names, "label_folders":[labels_folder_entry.get()]}
+    with open('train_data.yaml', 'w') as f:
+        yaml.dump(train_data, f, allow_unicode=True)
+
+
 def search_sources():
     source_folder = Path(source_folder_entry.get())
     sourcelist = [str(item.relative_to(source_folder)) for item in source_folder.glob(search_pattern_entry.get())]
@@ -147,11 +149,9 @@ def run_auto_label():
     sourcelist_yaml = {"videofiles": ["{VideoArchive}/"+item for item in source_list]}
     with open("videolist.yaml","w") as f:
         yaml.dump(sourcelist_yaml, f, allow_unicode=True)
-
-    with open("ann_manager_settings.json") as f:
-        settings = json.load(f)
         
     annotation_folder = Path(project_folder_entry.get()) / Path(annotation_subfolder_entry.get())
+    settings = ann_manager_settings
 
     settings["video_list_yaml"] = "videolist.yaml"
     settings["root_dataset_folder"] = str(annotation_folder)
@@ -220,6 +220,9 @@ if __name__ == "__main__":
     parser.add_argument('--source_folder', nargs='?', help='source folder name', default='')
     opt = parser.parse_args()
 
+    with open("ann_manager_settings.json") as f:
+        ann_manager_settings = json.load(f)
+
     workers_number = opt.workers_number
     if Path(opt.label_names_yaml).is_file():
         with open(opt.label_names_yaml) as f:
@@ -244,11 +247,6 @@ if __name__ == "__main__":
     root = Tk()
     root.geometry("840x625")
     root.resizable(False, False)
-
-    videolist_listbox = Listbox(width = 50)
-    processed_listbox = Listbox(width = 50)
-    videolist_listbox.grid(row=5, column=0, sticky=EW, padx=5, pady=5)
-    processed_listbox.grid(row=5, column=1, sticky=EW, padx=5, pady=5)
     
     source_folder = Path(os.path.expanduser(opt.source_folder))
 
@@ -256,20 +254,10 @@ if __name__ == "__main__":
     processed_list=[]
     videolist_listbox_fullnames=[]
 
-    ttk.Button(text="Process", command=process_video).grid(row=4, column=0, padx=5, pady=5, sticky=E)
-    ttk.Label(text="Annotation subfolder").grid(row=4, column=0, padx=5, pady=5, sticky=W)
-    annotation_subfolder_entry = ttk.Entry(width=7)
-    annotation_subfolder_entry.insert(0, "0")
-    annotation_subfolder_entry.grid(row=4, column=0, padx=150, pady=5, sticky=W)
-    ttk.Button(text="View", command=view_video).grid(row=4, column=1, padx=5, pady=5)
-    
     ttk.Label(text = "Project folder").grid(row=0, column=0, padx=5, pady=5, sticky=W)
     project_folder_entry = ttk.Entry(width=35)
     project_folder_entry.insert(0, os.path.expanduser(opt.project_folder[0]))
     project_folder_entry.grid(row=0, column=0, padx=5, pady=5, sticky=E)
-
-    ttk.Button(text="Generate images", command=generate_images).grid(row=6, column=0, padx=5, pady=5, sticky=W)
-    ttk.Button(text="Generate labels", command=generate_labels).grid(row=6, column=0, padx=130, pady=5, sticky=W)
 
     source_folder_entry = ttk.Entry(width=35)
     source_folder_entry.insert(0, source_folder)
@@ -281,23 +269,40 @@ if __name__ == "__main__":
     search_pattern_entry.grid(row=1, column=1, padx=5, pady=5, sticky=E)
     ttk.Label(text = "Search pattern").grid(row=1, column=1, padx=5, pady=5, sticky=W)
 
-    source_listbox = Listbox(width = 50)
-    source_listbox.grid(row=3, column=1, sticky=EW, padx=5, pady=5)
-
-    selected_listbox = Listbox(width = 50)
-    selected_listbox.grid(row=3, column=0, sticky=EW, padx=5, pady=5)
-
-    if 'selected' in save:
-        for item in save['selected']:
-            selected_listbox.insert(END, item)
-
     ttk.Button(text="Select", command=select_video).grid(row=2, column=1, padx=5, pady=5, sticky=W)
     ttk.Button(text="Select all", command=select_all_video).grid(row=2, column=1, padx=95, pady=5, sticky=W)
-    ttk.Button(text="Autolabel", command=run_auto_label).grid(row=4, column=0, padx=95, pady=5, sticky=E)
     
     ttk.Button(text="Remove", command=remove_item).grid(row=2, column=0, padx=5, pady=5, sticky=W)
     ttk.Button(text="Remove all", command=remove_all_items).grid(row=2, column=0, padx=95, pady=5, sticky=W)
     ttk.Button(text="Preview", command=video_preview).grid(row=2, column=0, padx=5, pady=5, sticky=E)
+
+    source_listbox = Listbox(width = 50)
+    source_listbox.grid(row=3, column=1, sticky=EW, padx=5, pady=5)
+    selected_listbox = Listbox(width = 50)
+    selected_listbox.grid(row=3, column=0, sticky=EW, padx=5, pady=5)
+
+    ttk.Button(text="Process", command=process_video).grid(row=4, column=0, padx=5, pady=5, sticky=E)
+    ttk.Label(text="Annotation subfolder").grid(row=4, column=0, padx=5, pady=5, sticky=W)
+    annotation_subfolder_entry = ttk.Entry(width=7)
+    annotation_subfolder_entry.insert(0, "0")
+    annotation_subfolder_entry.grid(row=4, column=0, padx=150, pady=5, sticky=W)
+    ttk.Button(text="View", command=view_video).grid(row=4, column=1, padx=5, pady=5)
+    ttk.Button(text="Autolabel", command=run_auto_label).grid(row=4, column=0, padx=95, pady=5, sticky=E)
+
+    videolist_listbox = Listbox(width = 50)
+    processed_listbox = Listbox(width = 50)
+    videolist_listbox.grid(row=5, column=0, sticky=EW, padx=5, pady=5)
+    processed_listbox.grid(row=5, column=1, sticky=EW, padx=5, pady=5)
+
+    ttk.Button(text="Generate dataset", command=generate_dataset).grid(row=6, column=1, padx=5, pady=5, sticky=E)
+    ttk.Label(text="Labels folder").grid(row=6, column=1, padx=5, pady=5, sticky=W)
+    labels_folder_entry = ttk.Entry(width=7)
+    labels_folder_entry.insert(0, "labels")
+    labels_folder_entry.grid(row=6, column=1, padx=100, pady=5, sticky=W)
+
+    if 'selected' in save:
+        for item in save['selected']:
+            selected_listbox.insert(END, item)
 
     search_pattern_entry.bind('<Return>', lambda event : search_sources())
     source_folder_entry.bind('<Return>', lambda event : search_sources())
